src/spoke1/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymysql==1.1.1
azure-storage-blob==12.24.0
